Move exceptions to subpackage
