+++ user-service/src/main/java/net/explorviz/security/server/main/DependencyInjectionBinder.java
@@ -7,7 +7,7 @@ import net.explorviz.security.services.BatchCreationService;
 import net.explorviz.security.services.RoleService;
 import net.explorviz.security.services.TokenService;
 import net.explorviz.security.services.UserService;
-import net.explorviz.security.services.UserValidationService;
+import net.explorviz.security.services.exceptions.UserValidationService;
 import net.explorviz.shared.common.idgen.IdGenerator;
 import net.explorviz.shared.common.injection.CommonDependencyInjectionBinder;
 import xyz.morphia.Datastore;
