+++ user-service/src/main/java/net/explorviz/security/server/main/SetupApplicationListener.java
@@ -5,8 +5,8 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.servlet.annotation.WebListener;
 import net.explorviz.security.services.RoleService;
-import net.explorviz.security.services.UserCrudException;
 import net.explorviz.security.services.UserService;
+import net.explorviz.security.services.exceptions.UserCrudException;
 import net.explorviz.security.util.PasswordStorage;
 import net.explorviz.security.util.PasswordStorage.CannotPerformOperationException;
 import net.explorviz.shared.security.model.User;
