+++ user-service/src/main/java/net/explorviz/security/server/resources/BatchRequestSubResource.java
@@ -12,9 +12,9 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import net.explorviz.security.model.UserBatchRequest;
 import net.explorviz.security.services.BatchCreationService;
-import net.explorviz.security.services.DuplicateUserException;
-import net.explorviz.security.services.MalformedBatchRequestException;
-import net.explorviz.security.services.UserCrudException;
+import net.explorviz.security.services.exceptions.DuplicateUserException;
+import net.explorviz.security.services.exceptions.MalformedBatchRequestException;
+import net.explorviz.security.services.exceptions.UserCrudException;
 import net.explorviz.shared.security.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
