+++ user-service/src/main/java/net/explorviz/security/services/exceptions/UserValidationService.java
@@ -1,8 +1,9 @@
-package net.explorviz.security.services;
+package net.explorviz.security.services.exceptions;
 
 import javax.inject.Inject;
 import javax.ws.rs.ForbiddenException;
 import net.explorviz.security.model.UserCredentials;
+import net.explorviz.security.services.UserService;
 import net.explorviz.security.util.PasswordStorage;
 import net.explorviz.security.util.PasswordStorage.CannotPerformOperationException;
 import net.explorviz.security.util.PasswordStorage.InvalidHashException;
