+++ user-service/src/test/java/net/explorviz/security/server/resources/UserResourceTest.java
@@ -16,8 +16,8 @@ import java.util.stream.Collectors;
 import javax.ws.rs.BadRequestException;
 import net.explorviz.security.server.resources.endpoints.UserResourceEndpointTest;
 import net.explorviz.security.services.RoleService;
-import net.explorviz.security.services.UserCrudException;
 import net.explorviz.security.services.UserService;
+import net.explorviz.security.services.exceptions.UserCrudException;
 import net.explorviz.security.util.PasswordStorage;
 import net.explorviz.security.util.PasswordStorage.CannotPerformOperationException;
 import net.explorviz.security.util.PasswordStorage.InvalidHashException;
