+++ user-service/src/test/java/net/explorviz/security/server/resources/endpoints/UserResourceEndpointTest.java
@@ -14,8 +14,8 @@ import javax.ws.rs.core.Response;
 import net.explorviz.security.server.main.DependencyInjectionBinder;
 import net.explorviz.security.server.resources.UserResource;
 import net.explorviz.security.services.RoleService;
-import net.explorviz.security.services.UserCrudException;
 import net.explorviz.security.services.UserService;
+import net.explorviz.security.services.exceptions.UserCrudException;
 import net.explorviz.security.testutils.TestDatasourceFactory;
 import net.explorviz.shared.exceptions.ErrorObjectHelper;
 import net.explorviz.shared.exceptions.JsonApiErrorObjectHelper;
